Replace incorrect writefile flag
A recent commit in the neovim codebase changed the behavior of using an
incorrect flag to writefile. Originally 'w' would fall through the other
flag checks and simply write to the file. Now it generates an unknown
flag error.

The correct flag is 's' which ensures the file is synced to disk which
is required for this plugin to work.

Fixes #5